v0.92 - More device control in csVisual crippled csDashboard
a) I've expanded the adhoc device control in python, and laid out a preliminary GUI design.

b) csDashboard in more functional, but the screen aspect is broken at this point. For now, csDashboard is mostly a relay for stepper control. I'm in the midst of overhaulling the serial communication between csBehavior and csDashboard, which will change the overall screen interface going forward.