v0.95 - Sinda Task Integration
a) Integrated Sinda's draft task, which I've titled "Trial Opto" I updated some of her edits to ensure the task conforms to the new GUI scheme. I made comments that start with cad: whenever I made a change. The task doesn't break the main program, so we will consider this the master draft. I will make a Sinda branch on the repository for future development, running code is encouraged to be pulled into the main branch.

b) Improved the GUI class. The GUI is essentially 100% independent of the tasks at this point, but not the other way around. I will work on this in the coming days, but for now two key things remain an issue to independence 1: I make the main GUI window in csGUI's init. The reason for this is that the tasks rely on components in that window. I will eventually make the main window creation a method of the class. 2: The tasks reference text variables in the GUI in many places, but they do not all fail gracefully. Most calls for GUI variables have a try/except scheme, but not all. I need to comb this. Moreover, I need to allow for updating these arguments via JSON or text etc.