v0.98 - Early No GUI Support
Added components to decouple the GUI from the tasks. You can open the script csBehavior.py with the path to a config file (an example "testConfig.txt" is included) and it will use variables from there to do the task. Ideally this replaces, or is replaced by the pandas csv, so I consider this an interim step.