Major Changes in Trial Variable Handling
Big overhaul in how trial-by-trial variables are handled.

All are specified in csVars using a catagortical dictionary and enumerating probabilities within those dictionaries.

In task, the variable dicts are looped over. Three main dicts: timing, sens, and optical

csParser will have to be overhauled as the attributes are now automatically generated and are different.

Also, I added a concept of local python time. Python was always keeping track of the teensy clock, as streamed over serial. But now, there are some local variables that are copies of that clock. I have safeguards in there now that prevent time from going backward, even for a sample. This is an equally big change, and more likley to solve or create performance issues.