better encapsulation of trial variables
Effecticvley everything about trial variables is encapsulated in the csVars class.

Added a function in csVars that allows for generating/regenerating variables on the fly (easily).

Got rid of the extra variable label dictionaries, and now have a key in the dicts whose value is the list.

Overall everything is much more efficient

I need to farm out the visual serial call, just like the analog updates.