important dictionary indexing changes
In some cases, I was asssuming the order in the dictionary was deterministic. Menaing that if I made a dictionary as vars={'a':1,'b':2,'c':3} then if you call for x in list ... then x would go a,b and then c, but this is not always the case. Dictionaries are efficient because they allow the OS to map them in memory as they please, this makes sense. On OSX, I've never seen it out of order, so I was luled into a false sense of security.

Now when trialVars are computed, they generate a list of computed numbers and an index. The index is just an ordered list of the key's in order of their processing. When they are used later, the index is refered to.